requirements.txt
appdirs==1.4.4
asgiref==3.6.0
autopep8==2.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
defusedxml==0.7.1
Django==3.2.7
django-autoslug==1.9.8
django-countries==7.2.1
django-filter==21.1
django-phonenumber-field==5.2.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
flake8==3.9.2
idna==3.4
inflection==0.5.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.6.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.0
phonenumbers==8.12.33
Pillow==8.3.2
psycopg2-binary==2.9.1
pycodestyle==2.10.0
pycparser==2.21
pyflakes==2.3.1
PyJWT==2.1.0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.7.1
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.14
